Avoid NoMethodError in datastreams controller
Connects to https://app.honeybadger.io/projects/49897/faults/48659064

This bug, which is not triggered very often, was caused by the upgrade to Rails 5. We adjusted the call in the events controller but not in the datastreams controller to avoid calling a method that is no longer defined on `ActionController::Parameters`.

As part of this work, reduce duplication of code by moving a now-identical
method into `ApplicationController`.