Allow for filtering which collections and items are exported
By convention, use a file called "collection_druids.txt", which is a plain-text file containing a number of bare druids, one per line. This file is produced by copying a column of druids out of a spreadsheet maintained by the PO. This will filter both the collections and the items that are exported.